refactor: move AllowedAlertSeverity type to components.ts and update imports across MIAlert and MISnackbarAlert
