--- src/types/MIAlert.ts
@@ -1,1 +1,0 @@
-export type AllowedAlertSeverity = 'success' | 'info' | 'warning' | 'error';
